Autoboxing etc
git-svn-id: file:///home/toad/git-migration/temprepository/branches/db4o/freenet@24471 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/client/async/SplitFileFetcherSubSegment.java
@@ -651,8 +651,8 @@ public class SplitFileFetcherSubSegment extends SendableGet implements SupportsB
 		int blockNo;
 		synchronized(segment) {
 			for(int i=0;i<blockNums.size();i++) {
-				Integer token = (Integer) blockNums.get(i);
-				int num = ((Integer)token).intValue();
+				Integer token = blockNums.get(i);
+				int num = (Integer)token;
 				Key k = segment.getBlockNodeKey(num, container);
 				if(k != null && k.equals(key)) {
 					blockNums.remove(i);
